+++ doc/whats_new.rst
@@ -306,6 +306,9 @@ Changelog
      with little observations attached to them, by `Gael Varoquaux`_.
 
 
+   - New transformer :ref:`neural_network.BernoulliRBM`
+     by `Yann Dauphin`_.
+
 API changes summary
 -------------------
    - Renamed all occurences of ``n_atoms`` to ``n_components`` for consistency.